--- src/app/orders/edit/[id]/loading.tsx
@@ -1,9 +1,0 @@
-import { Loader2 } from 'lucide-react'
-
-export default function Loading() {
-  return (
-    <h1 className="flex h-screen items-center justify-center">
-      <Loader2 className="h-10 w-10 animate-spin" />
-    </h1>
-  )
-}
