--- src/app/orders/page.tsx
@@ -1,10 +1,0 @@
-import { Wrap } from '@/components/comum/Wrap'
-import Layout from '../dashboard/layout'
-
-export default function Orders() {
-  return (
-    <Layout>
-      <Wrap>Pedidos</Wrap>
-    </Layout>
-  )
-}
