+++ src/template/pedidos/Table/columns.tsx
@@ -12,14 +12,18 @@ import {
 import { Button } from '@/components/ui/button'
 import { MoreHorizontal, Printer, SquareMenu, SquarePen, XCircle } from 'lucide-react'
 import { toast } from 'react-toastify'
+import { useFormContext } from 'react-hook-form'
+import { useModal } from '@/contexts/modal'
 import { Column } from '@/template/recheios/Table/columns/components'
+import { FormDataPedidos } from '@/app/pedidos/types'
 import { StatusProps } from '@/app/pedidos/schema'
 import { Badge, BadgeProps } from '@/components/ui/badge'
 import { useContextOrders } from '@/contexts/dataContexts/ordersContext/useContextOrders'
 import { useView } from '@/contexts/view'
 import { GetOrder } from '@/types/order'
 import { useModalPrint } from '@/contexts/modalPrint'
-import { useRouter } from 'next/navigation'
+import { api } from '@/services/api/apiClient'
+import { useQueryState } from 'nuqs'
 
 export const columns: ColumnDef<GetOrder>[] = [
   {
@@ -106,12 +110,74 @@ export const columns: ColumnDef<GetOrder>[] = [
     id: 'actions',
     cell: ({ row }) => {
       const { removeOrder, getAllOrders } = useContextOrders()
+      const methods = useFormContext<FormDataPedidos>()
       const linha = row.original
 
+      const { address, orderProduct, date, bolo, payment, docesPP, ...rest } = linha
+
+      const order: FormDataPedidos = {
+        address: address?.id,
+        value_frete: address?.value_frete,
+        logistic: address?.type_frete,
+        cakes: bolo?.map((cake) => ({
+          id: cake.id,
+          peso: cake.peso,
+          formato: cake.formato,
+          massa: cake.massa,
+          recheios: cake.recheio,
+          price: cake.price,
+          cobertura: cake.cobertura,
+          decoracao: cake.description ?? '',
+          banner: cake.banner,
+          tem_topper: cake.topper ? true : false,
+          topper: {
+            tema: cake.topper?.tema ?? '',
+            name: cake.topper?.name ?? '',
+            idade: cake.topper?.idade,
+            price: cake.topper?.price ?? 18,
+            description: cake.topper?.description ?? '',
+            banner: cake.topper?.banner ?? '',
+            recebido: cake.topper?.recebido ?? false,
+            fornecedor: cake.topper?.fornecedor,
+          },
+        })),
+        orderProduct: orderProduct.reduce(
+          (acc, item) => {
+            if (item.category.priority < 0) {
+              return acc
+            }
+            if (typeof acc[item.category.priority] === 'undefined') {
+              acc[item.category.priority] = [item]
+
+              return acc
+            }
+
+            acc[item.category.priority].push(item)
+
+            return acc
+          },
+          [] as FormDataPedidos['orderProduct'],
+        ),
+        docesPP,
+        date: new Date(date),
+        payment: payment.map((pay) => {
+          return {
+            date: pay.date ? new Date(pay.date) : new Date(),
+            paid: pay.paid,
+            value: pay.value,
+            formPayment: pay.type,
+          }
+        }),
+        ...rest,
+      }
+
+      const { handleOpenOrder } = useModal()
       const { handleOpen, setId } = useView()
       const { handleOpen: handleOpenPrint } = useModalPrint()
-      // const [, setClientName] = useQueryState('client')
-      const { push } = useRouter()
+      const [, setClientName] = useQueryState('client')
+      const [, setAddressComplete] = useQueryState('address_complete')
+      const [, setFreteCarro] = useQueryState('frete_carro')
+      const [, setFreteMoto] = useQueryState('frete_moto')
       return (
         <DropdownMenu>
           <DropdownMenuTrigger asChild>
@@ -146,7 +212,20 @@ export const columns: ColumnDef<GetOrder>[] = [
 
             <DropdownMenuSeparator />
 
-            <DropdownMenuItem onClick={async () => push(`/orders/edit/${linha.id}`)}>
+            <DropdownMenuItem
+              onClick={async () => {
+                const response = await api.get(`search-client/${linha.client.tel}`)
+                const client = response.data[0]
+
+                setClientName(client.name)
+                setFreteCarro(`${address?.frete_carro}`)
+                setFreteMoto(`${address?.frete_moto}`)
+
+                setAddressComplete(address?.address_complete)
+                methods.reset(order)
+                handleOpenOrder()
+              }}
+            >
               Editar
               <SquarePen className="ml-2 h-4 w-4" />
             </DropdownMenuItem>
